+++ Source/WebKit/Scripts/webkit/tests/TestSerializedType.serialization.in
@@ -168,3 +168,7 @@ webkit_platform_headers: "PlatformClass.h"
 [CustomHeader, WebKitPlatform] class WebKit::PlatformClass {
   int value;
 };
+
+[CustomEncoder] class WebKit::CustomEncoded {
+  int value;
+};
