+++ Source/WebKit/Shared/RemoteLayerTree/RemoteLayerTreeTransaction.h
@@ -64,57 +64,18 @@ class Encoder;
 
 namespace WebKit {
 
-enum class LayerChange : uint64_t {
-    NameChanged                         = 1LLU << 1,
-    ChildrenChanged                     = 1LLU << 2,
-    PositionChanged                     = 1LLU << 3,
-    BoundsChanged                       = 1LLU << 4,
-    BackgroundColorChanged              = 1LLU << 5,
-    AnchorPointChanged                  = 1LLU << 6,
-    BorderWidthChanged                  = 1LLU << 7,
-    BorderColorChanged                  = 1LLU << 8,
-    OpacityChanged                      = 1LLU << 9,
-    TransformChanged                    = 1LLU << 10,
-    SublayerTransformChanged            = 1LLU << 11,
-    HiddenChanged                       = 1LLU << 12,
-    GeometryFlippedChanged              = 1LLU << 13,
-    DoubleSidedChanged                  = 1LLU << 14,
-    MasksToBoundsChanged                = 1LLU << 15,
-    OpaqueChanged                       = 1LLU << 16,
-    ContentsHiddenChanged               = 1LLU << 17,
-    MaskLayerChanged                    = 1LLU << 18,
-    ClonedContentsChanged               = 1LLU << 19,
-    ContentsRectChanged                 = 1LLU << 20,
-    ContentsScaleChanged                = 1LLU << 21,
-    CornerRadiusChanged                 = 1LLU << 22,
-    ShapeRoundedRectChanged             = 1LLU << 23,
-    ShapePathChanged                    = 1LLU << 24,
-    MinificationFilterChanged           = 1LLU << 25,
-    MagnificationFilterChanged          = 1LLU << 26,
-    BlendModeChanged                    = 1LLU << 27,
-    WindRuleChanged                     = 1LLU << 28,
-    SpeedChanged                        = 1LLU << 29,
-    TimeOffsetChanged                   = 1LLU << 30,
-    BackingStoreChanged                 = 1LLU << 31,
-    BackingStoreAttachmentChanged       = 1LLU << 32,
-    FiltersChanged                      = 1LLU << 33,
-    AnimationsChanged                   = 1LLU << 34,
-    AntialiasesEdgesChanged             = 1LLU << 35,
-    CustomAppearanceChanged             = 1LLU << 36,
-    UserInteractionEnabledChanged       = 1LLU << 37,
-    EventRegionChanged                  = 1LLU << 38,
-    ScrollingNodeIDChanged              = 1LLU << 39,
-#if HAVE(CORE_ANIMATION_SEPARATED_LAYERS)
-    SeparatedChanged                    = 1LLU << 40,
-#if HAVE(CORE_ANIMATION_SEPARATED_PORTALS)
-    SeparatedPortalChanged              = 1LLU << 41,
-    DescendentOfSeparatedPortalChanged  = 1LLU << 42,
-#endif
-#endif
-    VideoGravityChanged                 = 1LLU << 44,
-#if ENABLE(INTERACTION_REGIONS_IN_EVENT_REGION)
-    CoverageRectChanged                 = 1LLU << 45,
-#endif
+struct LayerProperties;
+using LayerPropertiesMap = HashMap<WebCore::PlatformLayerIdentifier, UniqueRef<LayerProperties>>;
+
+struct ChangedLayers {
+    HashSet<Ref<PlatformCALayerRemote>> changedLayers; // Only used in the Web process.
+    LayerPropertiesMap changedLayerProperties; // Only used in the UI process.
+
+    ChangedLayers();
+    ChangedLayers(ChangedLayers&&);
+    ChangedLayers& operator=(ChangedLayers&&);
+    ChangedLayers(LayerPropertiesMap&&);
+    ~ChangedLayers();
 };
 
 class RemoteLayerTreeTransaction {
@@ -145,101 +106,10 @@ public:
 #endif
     };
 
-    struct LayerProperties {
-        WTF_MAKE_STRUCT_FAST_ALLOCATED;
-        LayerProperties();
-        LayerProperties(const LayerProperties& other);
-
-        void encode(IPC::Encoder&) const;
-        static WARN_UNUSED_RETURN bool decode(IPC::Decoder&, LayerProperties&);
-
-        void notePropertiesChanged(OptionSet<LayerChange> changeFlags)
-        {
-            changedProperties.add(changeFlags);
-            everChangedProperties.add(changeFlags);
-        }
-
-        void resetChangedProperties()
-        {
-            changedProperties = { };
-        }
-
-        OptionSet<LayerChange> changedProperties;
-        OptionSet<LayerChange> everChangedProperties;
-
-        String name;
-        std::unique_ptr<WebCore::TransformationMatrix> transform;
-        std::unique_ptr<WebCore::TransformationMatrix> sublayerTransform;
-        std::unique_ptr<WebCore::FloatRoundedRect> shapeRoundedRect;
-
-        Vector<WebCore::PlatformLayerIdentifier> children;
-
-        Vector<std::pair<String, PlatformCAAnimationRemote::Properties>> addedAnimations;
-        HashSet<String> keysOfAnimationsToRemove;
-#if ENABLE(THREADED_ANIMATION_RESOLUTION)
-        Vector<Ref<WebCore::AcceleratedEffect>> effects;
-        WebCore::AcceleratedEffectValues baseValues;
-#endif
-
-        WebCore::FloatPoint3D position;
-        WebCore::FloatPoint3D anchorPoint { 0.5, 0.5, 0 };
-        WebCore::FloatRect bounds;
-        WebCore::FloatRect contentsRect { 0, 0, 1, 1 };
-        // Used in the WebContent process.
-        std::unique_ptr<RemoteLayerBackingStore> backingStore;
-        // Used in the UI process.
-        std::unique_ptr<RemoteLayerBackingStoreProperties> backingStoreProperties;
-        std::unique_ptr<WebCore::FilterOperations> filters;
-        WebCore::Path shapePath;
-        Markable<WebCore::PlatformLayerIdentifier> maskLayerID;
-        Markable<WebCore::PlatformLayerIdentifier> clonedLayerID;
-#if ENABLE(SCROLLING_THREAD)
-        WebCore::ScrollingNodeID scrollingNodeID { 0 };
-#endif
-        double timeOffset { 0 };
-        float speed { 1 };
-        float contentsScale { 1 };
-        float cornerRadius { 0 };
-        float borderWidth { 0 };
-        float opacity { 1 };
-        WebCore::Color backgroundColor { WebCore::Color::transparentBlack };
-        WebCore::Color borderColor { WebCore::Color::black };
-        WebCore::GraphicsLayer::CustomAppearance customAppearance { WebCore::GraphicsLayer::CustomAppearance::None };
-        WebCore::PlatformCALayer::FilterType minificationFilter { WebCore::PlatformCALayer::FilterType::Linear };
-        WebCore::PlatformCALayer::FilterType magnificationFilter { WebCore::PlatformCALayer::FilterType::Linear };
-        WebCore::BlendMode blendMode { WebCore::BlendMode::Normal };
-        WebCore::WindRule windRule { WebCore::WindRule::NonZero };
-        WebCore::MediaPlayerVideoGravity videoGravity { WebCore::MediaPlayerVideoGravity::ResizeAspect };
-        bool antialiasesEdges { true };
-        bool hidden { false };
-        bool backingStoreAttached { true };
-        bool geometryFlipped { false };
-        bool doubleSided { false };
-        bool masksToBounds { false };
-        bool opaque { false };
-        bool contentsHidden { false };
-        bool userInteractionEnabled { true };
-        WebCore::EventRegion eventRegion;
-
-#if ENABLE(INTERACTION_REGIONS_IN_EVENT_REGION)
-        WebCore::FloatRect coverageRect;
-#endif
-#if HAVE(CORE_ANIMATION_SEPARATED_LAYERS)
-        bool isSeparated { false };
-#if HAVE(CORE_ANIMATION_SEPARATED_PORTALS)
-        bool isSeparatedPortal { false };
-        bool isDescendentOfSeparatedPortal { false };
-#endif
-#endif
-    };
-
     explicit RemoteLayerTreeTransaction();
     ~RemoteLayerTreeTransaction();
     RemoteLayerTreeTransaction(RemoteLayerTreeTransaction&&);
     RemoteLayerTreeTransaction& operator=(RemoteLayerTreeTransaction&&);
-
-    void encode(IPC::Encoder&) const;
-    static WARN_UNUSED_RETURN bool decode(IPC::Decoder&, RemoteLayerTreeTransaction&);
 
     WebCore::PlatformLayerIdentifier rootLayerID() const { return m_rootLayerID; }
     void setRootLayerID(WebCore::PlatformLayerIdentifier);
@@ -252,22 +122,17 @@ public:
     String description() const;
     void dump() const;
 #endif
-
-    typedef HashMap<WebCore::PlatformLayerIdentifier, std::unique_ptr<LayerProperties>> LayerPropertiesMap;
     
-    bool hasAnyLayerChanges() const
-    {
-        return m_changedLayerProperties.size() || m_changedLayers.size() || m_createdLayers.size() || m_destroyedLayerIDs.size() || m_layerIDsWithNewlyUnreachableBackingStore.size();
-    }
+    bool hasAnyLayerChanges() const;
 
     const Vector<LayerCreationProperties>& createdLayers() const { return m_createdLayers; }
     const Vector<WebCore::PlatformLayerIdentifier>& destroyedLayers() const { return m_destroyedLayerIDs; }
     const Vector<WebCore::PlatformLayerIdentifier>& layerIDsWithNewlyUnreachableBackingStore() const { return m_layerIDsWithNewlyUnreachableBackingStore; }
 
-    HashSet<RefPtr<PlatformCALayerRemote>>& changedLayers() { return m_changedLayers; }
+    HashSet<Ref<PlatformCALayerRemote>>& changedLayers();
 
-    const LayerPropertiesMap& changedLayerProperties() const { return m_changedLayerProperties; }
-    LayerPropertiesMap& changedLayerProperties() { return m_changedLayerProperties; }
+    const LayerPropertiesMap& changedLayerProperties() const;
+    LayerPropertiesMap& changedLayerProperties();
 
     void setRemoteContextHostedIdentifier(Markable<WebCore::LayerHostingContextIdentifier> identifier) { m_remoteContextHostedIdentifier = identifier; }
     Markable<WebCore::LayerHostingContextIdentifier> remoteContextHostedIdentifier() const { return m_remoteContextHostedIdentifier; }
@@ -367,9 +232,10 @@ public:
 #endif
 
 private:
+    friend struct IPC::ArgumentCoder<RemoteLayerTreeTransaction, void>;
+
     WebCore::PlatformLayerIdentifier m_rootLayerID;
-    HashSet<RefPtr<PlatformCALayerRemote>> m_changedLayers; // Only used in the Web process.
-    LayerPropertiesMap m_changedLayerProperties; // Only used in the UI process.
+    ChangedLayers m_changedLayers;
 
     Markable<WebCore::LayerHostingContextIdentifier> m_remoteContextHostedIdentifier;
